build the search tree.
was orginally using a BFS search when I realized I was sort of already
building  a tree in a BFS sort of way with how I was building the node
tree from leaves of the tree.

In this case, landing on the 'end topic' in the links list is the same as finding
the node in the tree later. So just follow that nodes parentage and there is your
wiki path.